+++ internal/api/application.go
@@ -46,9 +46,6 @@ func (a *ApplicationAPI) Create(ctx context.Context, req *pb.CreateApplicationRe
 		InstallationMargin: req.InstallationMargin,
 		OrganizationID:     req.OrganizationID,
 	}
-	if req.ChannelListID > 0 {
-		app.ChannelListID = &req.ChannelListID
-	}
 
 	if err := storage.CreateApplication(a.ctx.DB, &app); err != nil {
 		return nil, errToRPCError(err)
@@ -86,10 +83,6 @@ func (a *ApplicationAPI) Get(ctx context.Context, req *pb.GetApplicationRequest)
 		OrganizationID:     app.OrganizationID,
 	}
 
-	if app.ChannelListID != nil {
-		resp.ChannelListID = *app.ChannelListID
-	}
-
 	return &resp, nil
 }
 
@@ -118,11 +111,6 @@ func (a *ApplicationAPI) Update(ctx context.Context, req *pb.UpdateApplicationRe
 	app.ADRInterval = req.AdrInterval
 	app.InstallationMargin = req.InstallationMargin
 	app.OrganizationID = req.OrganizationID
-	if req.ChannelListID > 0 {
-		app.ChannelListID = &req.ChannelListID
-	} else {
-		app.ChannelListID = nil
-	}
 
 	err = storage.UpdateApplication(a.ctx.DB, app)
 	if err != nil {
@@ -226,10 +214,6 @@ func (a *ApplicationAPI) List(ctx context.Context, req *pb.ListApplicationReques
 			AdrInterval:        app.ADRInterval,
 			InstallationMargin: app.InstallationMargin,
 			OrganizationID:     app.OrganizationID,
-		}
-
-		if app.ChannelListID != nil {
-			item.ChannelListID = *app.ChannelListID
 		}
 
 		resp.Result = append(resp.Result, &item)