+++ docs/changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Development
+
+**Features & Changes:**
+
+* Channel-lists have been removed. Extra channels (if supported by the ISM
+  band) are now managed by LoRa Server configuration.
+
+**Note:** when upgrading to this version with `--db-automigrate` /
+`DB_AUTOMIGRATE` set, channel-list data will be removed.
+
+
 ## 0.8.0
 
 **Features & changes:**